Close #3 - Add a Code of Conduct
This commit adds a Code of Conduct that is a copy of
Contributor Covenant, a code of conduct for open source
projects distributed under a MIT License. The file that contains
the text of the code is called CODE_OF_CONDUCT.md. The email for
contact is set to "akronix5@gmail.com" and "akronix5@gmail.com"

A reference to the code of conduct and to its text is made in the
CONTRIBUTING.md file.