Test dataset (#13)
* Add a short dataset of GPS data.

In this commit a short dataset of different GPS product files is
added. Each of the files contains just 9 registers and not real
coordinates, in order to hide the exact position of the animals.

* Add a short dataset of GPS data.

In this commit a short dataset of different GPS product files is
added. Each of the files contains just 9 registers and not real
coordinates, in order to hide the exact position of the animals.